clean generate data, change graalvm GC config
