+++ src/main/resources/application.properties
@@ -49,6 +49,8 @@ spring.rabbitmq.virtual-host=/
 
 fixed-rate=500
 
+#generate-data
+
